Created factories to remove hardcoding
I created an ItemFactory that creates a concrete item from an itemBase, and created a decorator factory that creates an ItemDecorator from an sEnhancer. This way, we can create scriptable objects in the editor and add them to a recipelist. A crafter now only has to load the correct list.